Unify gRPC code into single .h/.cc files.
