+++ protocol_driver_allocator.cc
@@ -15,7 +15,6 @@
 #include "protocol_driver_allocator.h"
 
 #include "protocol_driver_grpc.h"
-#include "protocol_driver_grpc_async_callback.h"
 
 namespace distbench {
